fixed add to account for numbers < 0 and > -1
